序列化代码生成测试成功

+++ EmbedXrpcIdlParser/idltest1.cs
@@ -50,7 +50,7 @@ struct  Student_t
 
     [FieldIndex(6)]
     [MaxCount(IsFixed = true, MaxCount = 100, LenFieldIndex = 5)]
-    byte[] StudentId;
+    Int16[] StudentId;
 
     [FieldIndex(7)]
     Sex_t Sex;
